[refactor] move user to be under auth package
